GPU: fix DevicePassable nightly break + clean warnings (dev2026080206)
The nightly bump to 1.0.0b3.dev2026080206 broke GPU builds, invisible
under moon:check since GPU gates are release-preflight only:

- `layout` moved out of mojo into MAX — add the `max` pixi dependency
  (same split already handled for json/pixi.toml).
- Int/UInt no longer conform to DevicePassable and can't be passed to
  GPU kernels. Every kernel entry point (kernels/__init__.mojo,
  tensor_ops.mojo call sites, models/qwen.mojo + gemma*.mojo) now
  takes fixed-width Int32 args, shadow-cast back to Int inside the
  body so kernel logic is untouched.
- Span's `ptr=` kwarg renamed to `unsafe_ptr=` (blockcache.mojo).
- `f.seek()` now takes Int directly, not UInt64 (safetensors.mojo).
- memcpy/memset_zero rename follow-through: update the `from
  std.memory import` lines alongside the unsafe_memcpy call sites.
- Deprecated pointer-arithmetic/bitcast/getitem warnings across the
  GPU-only src tree + one dead local in qwen.mojo, cleaned the same
  way as the rest of the nightly-bump warning pass.

Verified: build-gpu-hello, test-kernels, test-simd-gemm, test-attention
all pass with real numerical results (max_abs errors in the 1e-6-1e-7
range against NumPy/reference), full server.mojo binary builds clean
with zero warnings, and the CPU `pixi run test` suite still passes.

+++ src/gpu_hello.mojo
@@ -30,7 +30,7 @@ def fma_kernel(
     a: TileTensor[dtype, type_of(layout), MutAnyOrigin],
     b: TileTensor[dtype, type_of(layout), MutAnyOrigin],
     c: TileTensor[dtype, type_of(layout), MutAnyOrigin],
-    size: Int,
+    size: UInt32,
 ):
     """Compute `c[i] = a[i] * b[i] + a[i]` elementwise over `size` elements.
 
@@ -41,7 +41,7 @@ def fma_kernel(
         size: Number of leading elements to process.
     """
     var tid = global_idx.x
-    if tid < size:
+    if tid < Int(size):
         c[tid] = a[tid] * b[tid] + a[tid]
 
 
@@ -75,7 +75,7 @@ def main() raises:
         a,
         b,
         c,
-        N,
+        UInt32(N),
         grid_dim=ceildiv(N, BLOCK),
         block_dim=BLOCK,
     )
